Use background color on bib cells
Now the empty bib cells in the timing recorder are marked red, so the user is suggested to fill them with a value.
Confirmed bibs are marked with a green background instead.
When the background of the bib cell is white, any entered digit will be added to that cell.